+++ packages/vue-components/stories/OmegaForm/Defaults.vue
@@ -114,6 +114,7 @@ const struct = {
   v: S.UndefinedOr(S.NonEmptyString)
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Effect Schema Class generic param
 class ClassSchema extends S.Class<ClassSchema, any>("ClassSchema")(struct) {}
 const schema = S.Struct(struct)