+++ packages/eslint-codegen-model/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @effect-app/eslint-codegen-model
 
+## 2.0.0-beta.11
+
+### Patch Changes
+
+- c1e73de:
+
+## 2.0.0-beta.10
+
+### Patch Changes
+
+- 6fff09c: unify encoded function for when you use encodedKeys
+
 ## 2.0.0-beta.9
 
 ### Patch Changes